pyproject.toml
[project]
name = "Ekalavya"
version = "0.1.0"
description = "An AI-based Personal Tutor for Every Child"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.2.0",
    "groq>=0.31.1",
    "gtts>=2.5.4",
    "joblib>=1.5.2",
    "opencv-contrib-python>=4.12.0.88",
    "opencv-python-headless>=4.11.0.86",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "pyqt6>=6.9.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.18.1",
]
